requirements.txt
gspread==5.12.0
requests==2.31.0
python-dotenv==1.0.0
oauth2client==4.1.3
